Fix deflating in examples/mlzlib/zpipe.ml
For the case that the size of the last compressed output overflows the
size of the chunk, so the gz file will be truncated in middle.

This would not happen if it flushes every time, but since Z_NO_FLUSH is
explicitly specified in zpipe.ml, zlib is expected to pool output.
If Z_SYNC_FLUSH is used instead, the compression ratio will be reduced.